+++ parseArgsCLI.js
@@ -1,0 +1,6 @@
+function getValue(flag) {
+  const flagIndex = process.argv.indexOf(flag);
+  return flagIndex !== -1 ? process.argv[flagIndex + 1] : null;
+}
+
+module.exports = getValue
